fix: package.json for raxui & fix calendar
