+++ packages/rax-picture/package.json
@@ -1,6 +1,6 @@
 {
   "name": "rax-picture",
-  "version": "0.1.9",
+  "version": "0.2.2",
   "description": "picture UI for Rax.",
   "license": "BSD-3-Clause",
   "main": "lib/index.js",
@@ -10,17 +10,24 @@
   },
   "keywords": [
     "Rax",
-    "rax-component",
-    "react-component"
+    "RaxUI"
   ],
   "bugs": {
     "url": "https://github.com/alibaba/rax/issues"
   },
   "homepage": "https://github.com/alibaba/rax#readme",
+  "engines": {
+    "npm": ">=3.0.0"
+  },
   "dependencies": {
-    "rax": "^0.2.2",
     "universal-env": "^0.2.2",
     "rax-image": "^0.2.2",
     "rax-view": "^0.2.2"
+  },
+  "peerDependencies": {
+    "rax": "0.x"
+  },
+  "devDependencies": {
+    "rax-test-renderer": "^0.2.2"
   }
 }
