+++ packages/rax-tabbar/package.json
@@ -1,6 +1,6 @@
 {
   "name": "rax-tabbar",
-  "version": "0.1.9",
+  "version": "0.2.2",
   "description": "tabbar UI for Rax.",
   "license": "BSD-3-Clause",
   "main": "lib/index.js",
@@ -10,19 +10,26 @@
   },
   "keywords": [
     "Rax",
-    "rax-component",
-    "react-component"
+    "RaxUI"
   ],
   "bugs": {
     "url": "https://github.com/alibaba/rax/issues"
   },
   "homepage": "https://github.com/alibaba/rax#readme",
+  "engines": {
+    "npm": ">=3.0.0"
+  },
   "dependencies": {
-    "rax": "^0.2.2",
     "rax-view": "^0.2.2",
     "rax-text": "^0.2.2",
     "rax-touchable": "^0.2.2",
     "rax-image": "^0.2.2",
     "rax-scrollview": "^0.2.2"
+  },
+  "peerDependencies": {
+    "rax": "0.x"
+  },
+  "devDependencies": {
+    "rax-test-renderer": "^0.2.2"
   }
 }
